Implemented Vector and Acceleration based movement with interpolation
also added autojoin and start game for faster testing

+++ src/main/java/game/stages/InLobby.java
@@ -174,6 +174,9 @@ public class InLobby {
   /** Deletes all the texts from this Page from the rendering list. */
   @SuppressWarnings("Duplicates")
   public static void done() {
+    if (lobbyname == null) {
+      return;
+    }
     lobbyname.delete();
     for (int i = 0; i < names.length; i++) {
       names[i].delete();
